pyproject.toml
[project]
name = "gmail-ai-agent"
version = "0.1.0"
description = "AI agent that connects to Gmail and searches emails using natural language"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-generativeai>=0.3.0",
    "cryptography>=41.0.0",
    "tabulate>=0.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gmail_agent"]
